requirements.txt
requests
python-dotenv
websocket-client
Pillow
gradio
opencv-python-headless
onnxruntime
numpy
